requirements.txt
# Orbis v2.0 - Core Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Data & Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg[binary]==3.2.3

# Cache & Queue
redis==5.0.8
celery[redis]==5.4.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
PyJWT==2.9.0
cryptography==43.0.3

# HTTP
httpx==0.27.2

# WebSocket
websockets==13.1

# Config
python-dotenv==1.0.1

# Monitoring
prometheus-client==0.21.0
structlog==24.4.0

# System utilities
psutil==6.0.0

# Email (for notifications)
aiosmtplib==3.0.2

# ML (Optional - install separately with requirements-ml.txt)
# See requirements-ml.txt for ML dependencies

# Advanced Analytics (first set)
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
scipy==1.13.0

# AI/ML Advanced
openai==1.12.0
transformers==4.37.2
torch==2.1.2
torchaudio==2.1.2

# Enterprise Features
python-ldap==3.4.3
saml2==4.9.0

# Mobile Development
fastapi-mobile==0.1.0

# Additional Dependencies for Billion-Dollar Scale
flower==2.0.1
gunicorn==21.2.0
uvloop==0.19.0
httptools==0.6.1

# Enterprise Database & Performance
asyncpg==0.29.0
psycopg2-binary==2.9.9
aioredis==2.0.1

# Military-Grade Encryption
argon2-cffi==23.1.0
pycryptodome==3.19.0

# Cloud Storage & Backup
boto3==1.34.34
botocore==1.34.34
awscli==1.32.34

# Compression & Archiving
# gzip, bz2, lzma are built-in Python modules

# Monitoring & Metrics
prometheus-client==0.19.0
grafana-api==1.0.3
elasticsearch==8.11.1
kibana==8.11.1

# Performance Optimization
numba==0.58.1
cython==3.0.6
# pypy is a separate Python implementation, not a package

# Security & Compliance
python-ldap==3.4.3
saml2==7.4.1
oauthlib==3.2.2
authlib==1.2.1

# Enterprise Features
schedule==1.2.0
APScheduler==3.10.4
python-multipart==0.0.6
email-validator==2.1.0
phonenumbers==8.13.27

# Additional Analytics libraries
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Real-time Processing
kafka-python==2.0.2
pika==1.3.2
pyzmq==25.1.2

# Testing & Quality
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
mkdocs==1.5.3
mkdocs-material==9.4.8

# BigTech Security Enhancements
psycopg2-binary==2.9.11
pyotp==2.9.0
qrcode==8.2

# Voice Cloning with Coqui TTS (works with 5+ seconds audio)
TTS==0.22.0
